feat(api): disallow delete of assigned

+++ apps/api/src/app/layouts/use-cases/delete-layout/delete-layout.use-case.ts
@@ -1,14 +1,19 @@
-import { Injectable, InternalServerErrorException } from '@nestjs/common';
+import { ConflictException, Injectable, InternalServerErrorException } from '@nestjs/common';
 import { LayoutEntity, LayoutRepository } from '@novu/dal';
 
 import { DeleteLayoutCommand } from './delete-layout.command';
 
 import { LayoutDto } from '../../dtos';
+import { CheckLayoutIsUsedCommand, CheckLayoutIsUsedUseCase } from '../check-layout-is-used';
 import { GetLayoutCommand, GetLayoutUseCase } from '../get-layout';
 
 @Injectable()
 export class DeleteLayoutUseCase {
-  constructor(private getLayoutUseCase: GetLayoutUseCase, private layoutRepository: LayoutRepository) {}
+  constructor(
+    private getLayoutUseCase: GetLayoutUseCase,
+    private checkLayoutIsUsed: CheckLayoutIsUsedUseCase,
+    private layoutRepository: LayoutRepository
+  ) {}
 
   async execute(command: DeleteLayoutCommand): Promise<void> {
     const getLayoutCommand = GetLayoutCommand.create({
@@ -16,6 +21,17 @@ export class DeleteLayoutUseCase {
     });
 
     const layout = await this.getLayoutUseCase.execute(getLayoutCommand);
+
+    const isUsed = await this.checkLayoutIsUsed.execute(
+      CheckLayoutIsUsedCommand.create({
+        environmentId: command.environmentId,
+        layoutId: command.layoutId,
+        organizationId: command.organizationId,
+      })
+    );
+    if (isUsed) {
+      throw new ConflictException(`Layout with id ${command.layoutId} is being used so it can not be deleted`);
+    }
 
     await this.layoutRepository.deleteLayout(
       LayoutRepository.convertStringToObjectId(command.layoutId),
